+++ tests/functional/adapter/materialized_view_tests/fixtures.py
@@ -64,3 +64,46 @@ models:
         relation: true
         columns: true
 """
+
+complex_types_materialized_view = """
+{{ config(
+    materialized='materialized_view',
+    schedule = {
+        'cron': '0 0 * * * ? *',
+        'time_zone': 'Etc/UTC'
+    },
+) }}
+SELECT
+  named_struct(
+    'field1', map('a', 1, 'b', 2),
+    'field2', array(10, 20, 30),
+    'field3', 3,
+    'field4', 4,
+    'field5', 5,
+    'field6', 6,
+    'field7', 7,
+    'field8', 8,
+    'field9', 9,
+    'field10', 10,
+    'field11', 11,
+    'field12', 12,
+    'field13', 13,
+    'field14', 14,
+    'field15', 15,
+    'field16', 16,
+    'field17', 17,
+    'field18', 18,
+    'field19', 19,
+    'field20', 20,
+    'field21', 21,
+    'field22', 22,
+    'field23', 23,
+    'field24', 24,
+    'field25', 25,
+    'field26', 26,
+    'field27', 27,
+    'field28', 28,
+    'field29', 29,
+    'field30', 30
+  ) AS my_struct;
+"""